HS-34-04: decompose intel.py into an intel/ package
The 1,066-line module becomes a package with a full re-export __init__,
mirroring the Phase-31 db split. No caller or test import changed.

Layering (acyclic): models (dataclasses + constants) <- parsing (JSON coercion)
/ providers (resolution + egress posture) <- engine (MeetingIntel).
build_configured_meeting_intel imports the engine lazily to break the
providers<->engine edge.

Monkeypatch hazards preserved with tests UNCHANGED:
- OpenAI / Llama: the optional-dependency import head stays in __init__ (the
  patch target); providers + engine read them via _intel_pkg.OpenAI/.Llama
  (and _intel_pkg._OPENAI_IMPORT_ERROR/._IMPORT_ERROR) at call time, so a
  package patch reaches both (egress-invariant + cloud + streaming tests).
- resolve_intel_provider: re-export makes the package attr patchable; the one
  intra-package caller (resolve_llm_capability) calls it via _intel_pkg so the
  plugin-host patch reaches it.

Decorator-aware AST extraction; relative imports bumped one dot; imports trimmed
via ruff --fix; F821 caught _IMPORT_ERROR / socket / a MeetingIntel forward-ref
(TYPE_CHECKING import). New tests/unit/test_intel_package.py guards the surface +
asserts the engine reads OpenAI through the package. Intel suite (89) + full
suite green 1962/15; package ruff + F821 clean.
