+++ holdspeak/intel/providers.py
@@ -1,0 +1,267 @@
+"""Intel provider resolution + egress posture (HS-34-04).
+
+`OpenAI`/`Llama` live in the package `__init__` (the optional-dependency import
+head) and are read here *via the package* (`_intel_pkg.OpenAI`/`.Llama`) so tests
+that monkeypatch `holdspeak.intel.OpenAI` / `holdspeak.intel.Llama` are honored.
+"""
+
+from __future__ import annotations
+
+import os
+from pathlib import Path
+from typing import TYPE_CHECKING, Any, Optional
+from urllib.parse import urlparse
+
+import holdspeak.intel as _intel_pkg
+
+from ..logging_config import get_logger
+
+if TYPE_CHECKING:
+    from .engine import MeetingIntel
+from .models import (
+    DEFAULT_INTEL_CLOUD_API_KEY_ENV,
+    DEFAULT_INTEL_CLOUD_MODEL,
+    DEFAULT_INTEL_MODEL_PATH,
+    DEFAULT_INTEL_PROVIDER,
+    SELF_HOSTED_CLOUD_API_KEY_PLACEHOLDER,
+    VALID_INTEL_PROVIDERS,
+)
+
+log = get_logger("intel")
+
+
+def _normalize_provider(provider: Optional[str]) -> str:
+    value = (provider or DEFAULT_INTEL_PROVIDER).strip().lower()
+    if value not in VALID_INTEL_PROVIDERS:
+        return DEFAULT_INTEL_PROVIDER
+    return value
+
+
+def _resolve_cloud_api_key(api_key_env: Optional[str]) -> Optional[str]:
+    env_name = (api_key_env or DEFAULT_INTEL_CLOUD_API_KEY_ENV).strip()
+    if not env_name:
+        env_name = DEFAULT_INTEL_CLOUD_API_KEY_ENV
+    value = os.environ.get(env_name)
+    if value:
+        return value.strip() or None
+    return None
+
+
+def _is_self_hosted_base_url(base_url: Optional[str]) -> bool:
+    """True when a custom (non-default) cloud base URL is configured."""
+    return bool(base_url and base_url.strip())
+
+
+def _effective_cloud_api_key(
+    api_key_env: Optional[str], base_url: Optional[str]
+) -> Optional[str]:
+    """Resolve the key to hand the OpenAI client.
+
+    Returns the env key when set. For a self-hosted endpoint (any custom
+    ``base_url``) with no key, returns a placeholder so the SDK can connect.
+    Returns ``None`` only when talking to the default OpenAI API with no key.
+    """
+    key = _resolve_cloud_api_key(api_key_env)
+    if key:
+        return key
+    if _is_self_hosted_base_url(base_url):
+        return SELF_HOSTED_CLOUD_API_KEY_PLACEHOLDER
+    return None
+
+
+def _validate_base_url(base_url: Optional[str]) -> Optional[str]:
+    if not base_url:
+        return None
+    value = base_url.strip()
+    if not value:
+        return None
+    parsed = urlparse(value)
+    if parsed.scheme not in {"http", "https"} or not parsed.netloc:
+        return f"Invalid cloud base URL: {value}"
+    return None
+
+
+def get_local_intel_runtime_status(
+    model_path: str = DEFAULT_INTEL_MODEL_PATH,
+) -> tuple[bool, Optional[str]]:
+    """Return whether local meeting intelligence can run right now."""
+    if _intel_pkg.Llama is None:
+        return False, "llama-cpp-python is not available"
+
+    resolved = Path(model_path).expanduser()
+    if not resolved.exists():
+        return False, f"Intel model not found: {resolved}"
+
+    return True, None
+
+
+def get_cloud_intel_runtime_status(
+    *,
+    cloud_model: str = DEFAULT_INTEL_CLOUD_MODEL,
+    cloud_api_key_env: str = DEFAULT_INTEL_CLOUD_API_KEY_ENV,
+    cloud_base_url: Optional[str] = None,
+) -> tuple[bool, Optional[str]]:
+    """Return whether cloud meeting intelligence can run right now."""
+    if _intel_pkg.OpenAI is None:
+        return False, "openai package is not available"
+
+    if not (cloud_model or "").strip():
+        return False, "Cloud intel model is not configured"
+
+    base_url_error = _validate_base_url(cloud_base_url)
+    if base_url_error is not None:
+        return False, base_url_error
+
+    # A custom base_url means a self-hosted server that ignores the key, so
+    # only the default OpenAI endpoint actually requires one.
+    if not _effective_cloud_api_key(cloud_api_key_env, cloud_base_url):
+        return False, f"Missing API key in ${cloud_api_key_env}"
+
+    return True, None
+
+
+def resolve_intel_provider(
+    provider: str = DEFAULT_INTEL_PROVIDER,
+    *,
+    model_path: str = DEFAULT_INTEL_MODEL_PATH,
+    cloud_model: str = DEFAULT_INTEL_CLOUD_MODEL,
+    cloud_api_key_env: str = DEFAULT_INTEL_CLOUD_API_KEY_ENV,
+    cloud_base_url: Optional[str] = None,
+) -> tuple[Optional[str], Optional[str]]:
+    """Resolve the active provider for this runtime.
+
+    Returns:
+        (provider, None) on success where provider is "local" or "cloud".
+        (None, reason) when unavailable.
+    """
+    normalized = _normalize_provider(provider)
+
+    if normalized == "local":
+        ok, reason = get_local_intel_runtime_status(model_path)
+        return ("local", None) if ok else (None, reason)
+
+    if normalized == "cloud":
+        ok, reason = get_cloud_intel_runtime_status(
+            cloud_model=cloud_model,
+            cloud_api_key_env=cloud_api_key_env,
+            cloud_base_url=cloud_base_url,
+        )
+        return ("cloud", None) if ok else (None, reason)
+
+    # auto = local-first fallback to cloud
+    local_ok, local_reason = get_local_intel_runtime_status(model_path)
+    if local_ok:
+        return "local", None
+
+    cloud_ok, cloud_reason = get_cloud_intel_runtime_status(
+        cloud_model=cloud_model,
+        cloud_api_key_env=cloud_api_key_env,
+        cloud_base_url=cloud_base_url,
+    )
+    if cloud_ok:
+        return "cloud", None
+
+    return (
+        None,
+        "Local intel unavailable"
+        f" ({local_reason}); cloud intel unavailable ({cloud_reason})",
+    )
+
+
+def resolve_llm_capability(meeting_config: Any) -> bool:
+    """Whether the ``"llm"`` plugin capability should be enabled.
+
+    True iff meeting intelligence is enabled in config *and* an intel provider
+    resolves (HS-16-02). The check is cheap — `resolve_intel_provider` only
+    inspects config + file existence, it does not warm a model. Any failure
+    (including a malformed config) is non-fatal and yields ``False`` so the host
+    is still constructed; LLM-backed plugins then cleanly block at execute time.
+    """
+    try:
+        if not bool(getattr(meeting_config, "intel_enabled", False)):
+            return False
+        provider = getattr(meeting_config, "intel_provider", None) or DEFAULT_INTEL_PROVIDER
+        kwargs: dict[str, Any] = {
+            "cloud_model": getattr(meeting_config, "intel_cloud_model", None) or DEFAULT_INTEL_CLOUD_MODEL,
+            "cloud_api_key_env": getattr(meeting_config, "intel_cloud_api_key_env", None)
+            or DEFAULT_INTEL_CLOUD_API_KEY_ENV,
+            "cloud_base_url": getattr(meeting_config, "intel_cloud_base_url", None),
+        }
+        model_path = getattr(meeting_config, "intel_realtime_model", None)
+        if model_path:
+            kwargs["model_path"] = model_path
+        resolved, _reason = _intel_pkg.resolve_intel_provider(provider, **kwargs)
+        return resolved is not None
+    except Exception:
+        return False
+
+
+def build_configured_meeting_intel() -> "MeetingIntel":
+    """Construct a `MeetingIntel` from the user's saved meeting config.
+
+    Built-in plugins (`mermaid_architecture`, `action_owner_enforcer`, …) call
+    this for their default intel provider so they honour the configured endpoint
+    (e.g. a self-hosted `.43` cloud base URL) instead of the bare `MeetingIntel()`
+    module defaults — which would otherwise ignore the user's provider entirely.
+    """
+    from ..config import Config
+    from .engine import MeetingIntel
+
+    meeting = Config.load().meeting
+    kwargs: dict[str, Any] = {
+        "provider": getattr(meeting, "intel_provider", DEFAULT_INTEL_PROVIDER),
+        "cloud_model": getattr(meeting, "intel_cloud_model", DEFAULT_INTEL_CLOUD_MODEL),
+        "cloud_api_key_env": getattr(meeting, "intel_cloud_api_key_env", DEFAULT_INTEL_CLOUD_API_KEY_ENV),
+        "cloud_base_url": getattr(meeting, "intel_cloud_base_url", None),
+        "cloud_reasoning_effort": getattr(meeting, "intel_cloud_reasoning_effort", None),
+        "cloud_store": bool(getattr(meeting, "intel_cloud_store", False)),
+    }
+    model_path = getattr(meeting, "intel_realtime_model", None)
+    if model_path:
+        kwargs["model_path"] = model_path
+    return MeetingIntel(**kwargs)
+
+
+def intel_egress_posture(provider: str = DEFAULT_INTEL_PROVIDER) -> tuple[bool, str]:
+    """Describe whether the configured provider can send transcripts off-machine.
+
+    This is a *static* description of intent from config — it answers "can this
+    setting transmit a transcript to the cloud?", not "is a model loaded right
+    now?". It is the single source of truth for the egress posture surfaced in
+    ``holdspeak doctor`` and the web runtime status (HS-25-01).
+
+    Returns ``(can_transmit_offmachine, human_description)``.
+    """
+    normalized = _normalize_provider(provider)
+    if normalized == "local":
+        return False, "Local only — transcripts never leave this machine."
+    if normalized == "cloud":
+        return True, "Cloud — transcripts are sent to the configured cloud endpoint."
+    # auto = local-first, but will fall back to the cloud when no local model is
+    # available, so the configuration *can* transmit off-machine.
+    return (
+        True,
+        "Auto — local first, but falls back to sending transcripts to the cloud "
+        "when no local model is available.",
+    )
+
+
+def get_intel_runtime_status(
+    model_path: str = DEFAULT_INTEL_MODEL_PATH,
+    *,
+    provider: str = DEFAULT_INTEL_PROVIDER,
+    cloud_model: str = DEFAULT_INTEL_CLOUD_MODEL,
+    cloud_api_key_env: str = DEFAULT_INTEL_CLOUD_API_KEY_ENV,
+    cloud_base_url: Optional[str] = None,
+) -> tuple[bool, Optional[str]]:
+    """Return whether the configured meeting-intel mode can run right now."""
+    active, reason = resolve_intel_provider(
+        provider,
+        model_path=model_path,
+        cloud_model=cloud_model,
+        cloud_api_key_env=cloud_api_key_env,
+        cloud_base_url=cloud_base_url,
+    )
+    if active is None:
+        return False, reason
+    return True, None
